+++ pkg/virtualization/vmdk/streamoptimized.go
@@ -366,7 +366,8 @@ func (v *StreamOptimizedImage) TranslateOffset(off int64) (int64, int64, error) 
 	if grainOffset == GTEEmpty {
 		return 0, 0, ErrDataNotPresent
 	}
-	if grainOffset == GTEZeroed && v.SparseExtentHeader.Flags&uint32(FlagUseZeroedGrainTableEntry) != 0 {
+	// Per VMDK spec, GTE=1 is a valid sector offset when FlagUseZeroedGrainTableEntry is not set
+	if grainOffset == GTEZeroed && v.SparseExtentHeader.Flags&FlagUseZeroedGrainTableEntry != 0 {
 		return 0, 0, ErrDataNotPresent
 	}
